add home blueprint
