+++ blueprints/__init__.py
@@ -1,0 +1,14 @@
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+
+app = Flask(__name__)
+app.config.from_object('config.BaseConfig')
+
+db = SQLAlchemy(app)
+questions_list = []
+from blueprints.users.view import users_blueprint
+from blueprints.home.view import home_blueprint
+
+# register our blueprints
+app.register_blueprint(users_blueprint)
+app.register_blueprint(home_blueprint)
