+++ blueprints/users/view.py
@@ -4,8 +4,8 @@
 
 from flask import flash, redirect, render_template, request, \
     session, url_for, Blueprint
-from app import app
 from functools import wraps
+from blueprints import db
 
 ################
 #### config ####
@@ -19,8 +19,6 @@ users_blueprint = Blueprint(
 ##########################
 #### helper functions ####
 ##########################
-
-
 def login_required(test):
     @wraps(test)
     def wrap(*args, **kwargs):
@@ -47,7 +45,7 @@ def login():
         else:
             session['logged_in'] = True
             flash('You were logged in.')
-            return redirect(url_for('home'))
+            return redirect(url_for('home.home'))
     return render_template('login.html', error=error)
 
 
@@ -56,4 +54,4 @@ def login():
 def logout():
     session.pop('logged_in', None)
     flash('You were logged out.')
-    return redirect(url_for('welcome'))
+    return redirect(url_for('home.welcome'))
