+++ manage.py
@@ -1,7 +1,7 @@
 from flask_script import Manager
 from flask_migrate import Migrate, MigrateCommand
 
-from app import app, db
+from blueprints import app,db
 
 migrate = Migrate(app, db)
 manager = Manager(app)
